test/reftest/vendor/webkit/fast/dom/background-shorthand-csstext-expected.txt

background-shorthand-csstext.html:13 PASS: document.body.style.background == 'green' should be true and is.
background-shorthand-csstext.html:13 PASS: document.getElementById('div1').style.background == 'repeat-x white' should be true and is.
background-shorthand-csstext.html:13 PASS: document.getElementById('div2').style.background == 'repeat blue' should be true and is.
background-shorthand-csstext.html:13 PASS: document.getElementById('div3').style.background == 'none repeat rgb(255, 255, 255)' should be true and is.
